backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-multipart
python-dotenv
bcrypt
email-validator
PyJWT